Add functionality to ignore refresh session popup during an app session
Added back press handling for Android 13, Android 12, and lower. Implemented check for session refresh on app start.

Due to a change in Android 12 behavior (https://is.gd/7HzORm), the onCreate() and onDestroy() activity lifecycle callbacks will not be triggered anymore while using back navigation. Hence, onBackPressed() has to be overridden to detect if the app session is going to be closed by the user using back navigation.

In Android 13+, onBackPressed() will not be invoked. Hence, OnBackInvokedCallback is used to detect back navigation by the user.