refactor: improve code organization for OkHttp interceptors, Retrofit APIs and Hilt modules
This commit includes the addition of a network module and updates to the retrofit module. The newly added network module includes the configuration for OkHttpClient and different deserializers. Some of the deserializers added include Moshi, Gson, and JacksonConverterFactory for YAML files.

In addition to this, the retrofit module has been updated to include different APIs such as CleanApkRetrofit, CleanApkAppDetailsRetrofit, ExodusTrackerApi, FdroidApiInterface, EcloudApiInterface, AgeGroupApi, and FDroidMonitorApi.

Furthermore, there are also some changes in the directory structure of the codebase. The ageRating package has been renamed to parentalcontrol, and the RetrofitModule has been moved to the network package.

Lastly, the commit includes the addition of the okhttp3 logging interceptor in the build.gradle file.