refactor: modify dummy search behaviour
While searching for an app in App Lounge, initially we'll be using a dummy search for Facebook in the background in parallel with the original search to probe Google Play. We will expect both Facebook and Messenger apps to be returned in the search results. If it doesn't happen, then we'll refresh the token and try to search again.

Besides, if the token used in App Lounge gets refreshed in the process (or in any other scenario), we'll show a Toast to the user in debug builds.

Finally, HTTP logging level is set to request and headers only to avoid overflowing Logcat.