feat: block installation of third-party store apps when parental control is active
Similar to 'not-working' and 'zero-privacy' apps, list of third-party app stores, stored in the JSON at GitLab repo, are downloaded and saved into filesystem cache.

When trying to install such app, App Lounge checks if the package is already in the blocklist.

If any such package is already installed on the device prior to activating parental control, the package is checked against the blocklist and the data is sent to Parental Control app via AgeRatingProvider. Thus, it blocks the app from running.