feat: block installation of third-party store apps when parental control is active
Similar to 'not-working' and 'zero-privacy' apps, a list of third-party app stores, stored in the [JSON](https://gitlab.e.foundation/e/os/blocklist-app-lounge/-/blob/main/app-lounge-warning-list.json#L91) at GitLab repo, are downloaded and saved into filesystem cache.

When trying to install such an app, App Lounge checks if the package is already in the blocklist.

If any such package is already installed on the device prior to activating parental control, the package is checked against the blocklist and the data is sent to the Parental Control app via the AgeRatingProvider. Thus, parental control blocks the app from running.