fix: fix store visibility when they are checked/unchecked on Settings
Previously on commit df2ac320, Stores were introduced as an abstraction on Open Source, PWA and Google Play apps. However, when retrieving these stores, there was no check on the Stores class if the store is marked as selected by the user on the Settings screen. Every time App Lounge was closed and opened fresh, the stores' preference were loaded from the default — having all three checked.

The problem is visible when the checkboxes on the Settings screen are checked/unchecked and then App Lounge is started from Launcher by clearing it from the Recent drawer.

This fix takes the selection into consideration using AppLoungePreference and only returns those stores which are currently selected.